# Refactor: Simplify codebase with zero dependencies and comprehensive tests