--- a/core/__init__.py
+++ b/core/__init__.py
@@ -1,0 +1,7 @@
+"""
+Simplified Anki Dictionary Core Module
+Minimal dependencies, clean architecture
+"""
+
+__version__ = "2.0.0"
+__author__ = "Anki Dictionary Team"