… .openhands/microagents/repo.md
… README_REFACTORED.md
… REFACTORING_SUMMARY.md
… core/image_search.py
… demo_simplified.py
… run_tests.py
… standalone_analysis.md
… standalone_dict_demo.py
… tests/test_database.py
… tests/test_dictionary.py
… tests/test_image_search.py